Prevent astro:content from depending on Node builtins
