+++ .changeset/four-planets-smoke.md
@@ -1,0 +1,5 @@
+---
+'astro': patch
+---
+
+Prevent astro:content from depending on Node builtins
