+++ packages/astro/src/content/template/virtual-mod-assets.mjs
@@ -1,0 +1,9 @@
+import {
+	createImage
+} from 'astro/content/runtime-assets';
+
+const assetsDir = '@@ASSETS_DIR@@';
+
+export const image = createImage({
+	assetsDir,
+});
