+++ packages/astro/src/content/template/virtual-mod.mjs
@@ -3,8 +3,7 @@ import {
 	createCollectionToGlobResultMap,
 	createGetCollection,
 	createGetEntryBySlug,
-	createImage,
-} from 'astro/content/internal';
+} from 'astro/content/runtime';
 
 export { z } from 'astro/zod';
 
@@ -13,7 +12,6 @@ export function defineCollection(config) {
 }
 
 const contentDir = '@@CONTENT_DIR@@';
-const assetsDir = '@@ASSETS_DIR@@';
 
 const entryGlob = import.meta.glob('@@ENTRY_GLOB_PATH@@', {
 	query: { astroContent: true },
@@ -39,8 +37,4 @@ export const getCollection = createGetCollection({
 export const getEntryBySlug = createGetEntryBySlug({
 	getCollection,
 	collectionToRenderEntryMap,
-});
-
-export const image = createImage({
-	assetsDir,
 });
